# backend/openapi-app.yaml
# openapi-app.yaml
openapi: 3.0.0
info:
  title: Tasks Management API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /tasks:
    get:
      summary: Retrieve all tasks
      responses:
        "200":
          description: A JSON array of task objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "500":
          description: Internal Server Error
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad Request (Invalid Content)
        "500":
          description: Internal Server Error

  /tasks/{id}:
    get:
      summary: Retrieve a single task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A JSON object representing the task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not Found (Task not found)
        "500":
          description: Internal Server Error

    delete:
      summary: Delete a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Task deleted successfully
        "404":
          description: Not Found (Task not found)
        "500":
          description: Internal Server Error

    patch:
      summary: Update the status of a task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: integer
              required:
                - status
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad Request (Invalid Status)
        "404":
          description: Not Found (Task not found)
        "500":
          description: Internal Server Error

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        status:
          type: integer
          enum: [0, 1]
        completed_at:
          nullable: true
          type: string
          format: date-time